@asterite 2019-07-27
Sorry about the constant rebase, next time I'll keep adding commits to make it easier to review and just at the end I'll rebase everything before merging (I learned `rebase -i` and now I can't stop, hehe :-P)