@straight-shoota 2019-07-27
`git commit --fixup`  is great for this. It annotates a commit as fixup of a previous one and with `git rebase -i --autosquash` they get automatically inserted at the right place.