@asterite 2019-07-29
By the way, I benchmarked the [simple http server](https://github.com/crystal-lang/crystal/blob/master/samples/http_server.cr) on the samples directory before and after this change with `ab`. It's just a silly example but it can show whether this change had an effect on the overall http roundtrip.

Doing:

```
ab -k -c 100 -n 200000 127.0.0.1:8080/
```

Before:

```
Time taken for tests:   1.918 seconds
Requests per second:    104299.05 [#/sec] (mean)
Time per request:       0.959 [ms] (mean)
Time per request:       0.010 [ms] (mean, across all concurrent requests)
Transfer rate:          10287.31 [Kbytes/sec] received
```

After:

```
Time taken for tests:   1.716 seconds
Requests per second:    116551.95 [#/sec] (mean)
Time per request:       0.858 [ms] (mean)
Time per request:       0.009 [ms] (mean, across all concurrent requests)
Transfer rate:          11495.85 [Kbytes/sec] received
```

And I'm almost sure if we change Hash to have an open addressing implementation it could go up to `121750.26` requests per second (just a number out of the top of my head 😉).

By comparison, doing the same benchmark against a [simple server in Go](https://gowebexamples.com/http-server/) gives these results:

```
Time taken for tests:   2.767 seconds
Requests per second:    72285.75 [#/sec] (mean)
Time per request:       1.383 [ms] (mean)
Time per request:       0.014 [ms] (mean, across all concurrent requests)
Transfer rate:          10800.51 [Kbytes/sec] received
```

However, Go handles parallelism and when we'll have parallelism the performance will get a bit worse, but on the other hand a single request doing expensive CPU won't be able to stop the server from receiving other requests.