@RX14 2019-07-29
@asterite could you test with `wrk` instead in the future? It gives much better and more realistic results, and is the industry standard http benchmarking tool these days.