@RX14 2019-07-29
@asterite is that Go with `GOMAXPROCS` or is go actually using all cores and still losing?