@asterite 2019-07-29
> @asterite is that Go with GOMAXPROCS=1 or is go actually using all cores and still losing?

No, that's without specifying `GOMAXPROCS` (so I guess using all cores).

If I pass `GOMAXPROCS=1` I get:

```
Running 30s test @ http://127.0.0.1:8080/
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    18.14ms   88.69ms   2.00s    98.73%
    Req/Sec     3.14k   339.80     8.62k    85.96%
  1127120 requests in 30.10s, 138.66MB read
  Socket errors: connect 0, read 397, write 0, timeout 118
Requests/sec:  37442.65
Transfer/sec:      4.61MB
```

So I guess if with parallelism we can be close to Go numbers it'll be more than good ( /cc @waj @bcardiff )

> Also, are you going to address the review?

Yes, but in a couple of hours.